mengubah large class
