+++ src/lib/Person.java
@@ -1,0 +1,10 @@
+package lib;
+
+public class Person {
+    private String spouseName;
+	private String spouseIdNumber;
+
+    public void setSpouse(String spouseName, String spouseIdNumber) {
+		this.spouseName = spouseName;
+		this.spouseIdNumber = idNumber;
+}
